refactor: colocate logprob harness tooling and docs
